refactor(web): consolidate i18n under apps/web/i18n
