+++ apps/web/i18n/project.inlang/settings.json
