+++ apps/web/src/components/LocaleSwitcher.tsx
@@ -1,8 +1,8 @@
 // Locale switcher refs:
 // - Paraglide docs: https://inlang.com/m/gerre34r/library-inlang-paraglideJs
 // - Router example: https://github.com/TanStack/router/tree/main/examples/react/i18n-paraglide#switching-locale
-import { getLocale, locales, setLocale } from '#/paraglide/runtime'
-import { m } from '#/paraglide/messages'
+import { getLocale, locales, setLocale } from '#/i18n/runtime'
+import { m } from '#/i18n/messages'
 
 export default function ParaglideLocaleSwitcher() {
   const currentLocale = getLocale()
