+++ apps/web/src/i18n/messages.ts
@@ -1,0 +1,1 @@
+export * from '../../i18n/paraglide/messages.js'
