+++ apps/web/src/i18n/runtime.ts
@@ -1,0 +1,1 @@
+export * from '../../i18n/paraglide/runtime.js'
